Move get_next_view in utils to reuse it
